load data from Parse sever table Product

+++ PR392/app/src/main/java/com/example/lab10/Model/Product.java
@@ -1,19 +1,19 @@
 package com.example.lab10.Model;
 
 public class Product {
-    private double price;
+    private long price;
     private String name;
     private int number;
     private String description;
 
-    public Product(double price, String name, int number, String description) {
+    public Product(long price, String name, int number, String description) {
         this.price = price;
         this.name = name;
         this.number = number;
         this.description = description;
     }
 
-    public double getPrice() {
+    public long getPrice() {
         return price;
     }
